Added 128x64 OLED display
To keep program all inclusive included all files for displays instead of
library dependencies. Part of the reason is the required changes to the
libraries to make them work properly. Uses a software I2C library which
was modified so that the creation could be managed at compile time, not
run time. The adafruit display library was modified to use the software
wire library, and accept which pins are being used for software I2C
object at construction. The program source code is a little messy right
now since this was quick support for the display.